Added fetch()

+++ include/myARMSim.h
@@ -46,6 +46,6 @@ void mem();
 void write_back();
 
 
-int read_word(char *mem, unsigned long int address);
+unsigned long int read_word(char *mem, unsigned long int address);
 void write_word(char *mem, unsigned long int address, unsigned long int data);
